requirements.txt
asgiref==3.5.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
factory-boy==3.2.1
Faker==12.3.3
faker-food==0.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
packaging==23.0
psycopg2==2.9.5
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
